refactor(wallet): share path/address config helpers via cfgutil
oystercli was carrying its own normalizeAddress and cleanAndExpandPath,
duplicating logic that already exists for oyster. Move the ~/$VAR path
expansion into cfgutil.CleanAndExpandPath (oyster's config.go now
delegates to it, dropping its ~45-line copy) and switch oystercli to
cfgutil.NormalizeAddress. Both binaries share one implementation via the
wallet-internal cfgutil package they already use.
